requirements.txt
flask
flask-cors
pillow
transformers
torch
torchvision
gtts
pydub
werkzeug
